后端/FileSystem/requirements.txt
cryptography~=38.0.4
Flask~=2.2.2
Authlib~=1.2.0
Pillow~=9.3.0
APScheduler~=3.10.1
Werkzeug~=2.2.2
openpyxl~=3.1.2
SQLAlchemy~=1.4.44